[S2.6] Fix CI: cast bitfield reads in EXPECT_EQ to int (sign-compare)
build-32bit and build-macos-latest fail with -Werror=sign-compare:

  test_compression_workers.cpp:874:5: error: comparison of integers
    of different signs: 'const int' and 'const unsigned int'
    EXPECT_EQ(OBJ_ENCODING_RAW, u->encoding);
    EXPECT_EQ(OBJ_STRING, u->type);

robj's `encoding` and `type` are 4-bit unsigned bitfields (server.h);
they get integer-promoted to `unsigned int` when read. The OBJ_*
macros are signed int literals (e.g. OBJ_ENCODING_RAW == 0). gtest's
EXPECT_EQ instantiates CmpHelperEQ<int, unsigned int>, and the
mixed-sign `lhs == rhs` inside that helper trips -Wsign-compare /
-Werror.

Same root cause as the explicit `(int)u->refcount` cast on the next
line (OBJ_STATIC_REFCOUNT is also a signed-int macro and the
refcount field is unsigned). Apply the same cast to encoding and
type for consistency.

The 64-bit Linux gcc cell happens to type-promote without the
warning, which is why x86_64 passed and 32-bit + macOS clang did not.

The third failing cell (test-ubuntu-latest-compatibility, 8.1.4) is
unrelated — wget on download.valkey.io got HTTP 403 fetching the
8.1.4 release tarball. CDN/server transient issue, not a code issue.

Verified locally:
  - make -j2 -C src              → clean
  - ./runtest --single unit/type/compression  → 10/10 pass

+++ src/unit/test_compression_workers.cpp
@@ -871,8 +871,13 @@ TEST_F(CompressionWorkersTest, DecoderRoundTripsEncoder) {
 
     ASSERT_NE(u, nullptr);
     EXPECT_EQ(u, &view); /* helper returned the caller's stack robj */
-    EXPECT_EQ(OBJ_ENCODING_RAW, u->encoding);
-    EXPECT_EQ(OBJ_STRING, u->type);
+    /* Casts to int for the bitfield reads: u->encoding, u->type, and
+     * u->refcount are unsigned bitfields (server.h §robj). The OBJ_*
+     * macros are signed int literals (e.g. OBJ_ENCODING_RAW == 0).
+     * gcc/clang -Wsign-compare with -Werror flags the mixed-sign
+     * comparison; the cast aligns both sides to int. */
+    EXPECT_EQ(OBJ_ENCODING_RAW, (int)u->encoding);
+    EXPECT_EQ(OBJ_STRING, (int)u->type);
     EXPECT_EQ(OBJ_STATIC_REFCOUNT, (int)u->refcount);
 
     sds u_sds = (sds)objectGetVal(u);
